Extract LarkyByteArray::Elems to a separate class called LarkyByteIterable
